Move enum LibraryItemType
