--- src/app/library-css/constants/library-item-type.ts
@@ -1,5 +1,0 @@
-export enum LibraryItemType {
-  html = 0,
-  ngBootstrap = 1,
-  fhiAngular = 2
-}
