+++ src/mock-db-data/library-css/components/accordion.ts
@@ -1,7 +1,6 @@
-import { LibraryItem } from 'src/app/shared/models/library-item.model';
+import { LibraryItem, LibraryItemType } from 'src/app/shared/models/library-item.model';
 import { LibraryMenuLevel2Categories } from 'src/app/_common/constants/library-menu-level-2-categories';
 import { BootstrapBaseUrl, NgBootstrapBaseUrl } from 'src/app/library-css/constants/bootstrap-base-urls';
-import { LibraryItemType } from 'src/app/library-css/constants/library-item-type';
 
 const categoryNames = LibraryMenuLevel2Categories.cssComponentCategories;
